requirements.txt
bittensor
torch
pytesseract
pandas
faker
scipy
Pillow
PyMuPDF
reportlab
pdf2image
editdistance